backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sefa Sinanoğlu <sefasnngl@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = ">=0.116.0,<0.117.0"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
pydantic = {extras = ["email"], version = ">=2.11.7,<3.0.0"}
python-dotenv = ">=1.0.0,<2.0.0"
firebase-admin = ">=6.9.0,<7.0.0"
google-generativeai = ">=0.5.0,<1.0.0"


torch = "*"
torchvision = "*"
numpy = "*"
pillow = "*"
matplotlib = "*"
pyyaml = "*"
torchmetrics = "^1.7.4"
mlflow = "^3.1.4"
email-validator = "^2.2.0"
python-multipart = "^0.0.20"


[tool.poetry.scripts]
start = "uvicorn app.main:app --reload --host 0.0.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"